refactoring file names and folders
